+++ taichi/backends/opengl/opengl_program.h
@@ -2,8 +2,8 @@
 
 #include "taichi/backends/opengl/struct_opengl.h"
 
-#include "taichi/backends/opengl/opengl_kernel_launcher.h"
-#include "taichi/backends/opengl/opengl_api.h"
+#include "taichi/runtime/opengl/opengl_kernel_launcher.h"
+#include "taichi/runtime/opengl/opengl_api.h"
 #include "taichi/backends/opengl/codegen_opengl.h"
 
 #include "taichi/system/memory_pool.h"
