+++ taichi/program/extension.cpp
@@ -1,5 +1,4 @@
 #include "extension.h"
-//#include "taichi/backends/opengl/opengl_api.h"
 
 #include <unordered_map>
 #include <unordered_set>
