+++ taichi/python/export_misc.cpp
@@ -4,7 +4,7 @@
 *******************************************************************************/
 
 #include "taichi/backends/metal/api.h"
-#include "taichi/backends/opengl/opengl_api.h"
+#include "taichi/runtime/opengl/opengl_api.h"
 #include "taichi/runtime/vulkan/runtime.h"
 #include "taichi/backends/dx/dx_api.h"
 #include "taichi/common/core.h"
